backend/requirements.txt
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
manuf==1.1.5
openpyxl==3.1.5
packaging==26.0
pefile==2024.8.26
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
python-dotenv==1.2.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
scapy==2.7.0
setuptools==82.0.1
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
